Bazel-friendly elsint config setup
