--- libs/eslint-plugin-vx/.eslintignore
@@ -1,3 +1,0 @@
-/build
-/coverage
-/tests/fixtures
