+++ libs/eslint-plugin-vx/src/rules/BUILD.bazel
@@ -1,0 +1,13 @@
+load("//tools/ts_build:ts_library.bzl", "ts_library")
+
+ts_library(
+    name = "rules",
+    deps = [
+        "//:node_modules/@types/eslint",
+        "//:node_modules/@types/node",
+        "//:node_modules/@typescript-eslint/utils",
+        "//:node_modules/eslint",
+        "//:node_modules/typescript",
+        "//libs/eslint-plugin-vx/src/util",
+    ],
+)
