+++ libs/eslint-plugin-vx/src/rules/index.ts
@@ -5,7 +5,6 @@ import gtsConstants from './gts_constants';
 import gtsDirectModuleExportAccessOnly from './gts_direct_module_export_access_only';
 import gtsFuncStyle from './gts_func_style';
 import gtsIdentifiers from './gts_identifiers';
-import gtsJsdoc from './gts_jsdoc';
 import gtsModuleSnakeCase from './gts_module_snake_case';
 import gtsNoArrayConstructor from './gts_no_array_constructor';
 import gtsNoConstEnum from './gts_no_const_enum';
@@ -16,7 +15,6 @@ import gtsNoImportExportType from './gts_no_import_export_type';
 import gtsNoPrivateFields from './gts_no_private_fields';
 import gtsNoPublicClassFields from './gts_no_public_class_fields';
 import gtsNoPublicModifier from './gts_no_public_modifier';
-import gtsNoReturnTypeOnlyGenerics from './gts_no_return_type_only_generics';
 import gtsNoUnnecessaryHasOwnPropertyCheck from './gts_no_unnecessary_has_own_property_check';
 import gtsObjectLiteralTypes from './gts_object_literal_types';
 import gtsParameterProperties from './gts_parameter_properties';
@@ -28,7 +26,6 @@ import gtsUseOptionals from './gts_use_optionals';
 import noArraySortMutation from './no_array_sort_mutation';
 import noAssertStringOrNumber from './no_assert_truthiness';
 import noFloatingVoids from './no_floating_results';
-import noImportSubfolders from './no_import_workspace_subfolders';
 import noJestToBe from './no_jest_to_be';
 import noReactHookMutationDependency from './no_react_hook_mutation_dependency';
 
@@ -38,7 +35,6 @@ const rules: Record<string, Rule.RuleModule> = {
   'gts-direct-module-export-access-only': gtsDirectModuleExportAccessOnly,
   'gts-func-style': gtsFuncStyle,
   'gts-identifiers': gtsIdentifiers,
-  'gts-jsdoc': gtsJsdoc,
   'gts-module-snake-case': gtsModuleSnakeCase,
   'gts-no-array-constructor': gtsNoArrayConstructor,
   'gts-no-const-enum': gtsNoConstEnum,
@@ -49,7 +45,6 @@ const rules: Record<string, Rule.RuleModule> = {
   'gts-no-private-fields': gtsNoPrivateFields,
   'gts-no-public-class-fields': gtsNoPublicClassFields,
   'gts-no-public-modifier': gtsNoPublicModifier,
-  'gts-no-return-type-only-generics': gtsNoReturnTypeOnlyGenerics,
   'gts-no-unnecessary-has-own-property-check':
     gtsNoUnnecessaryHasOwnPropertyCheck,
   'gts-object-literal-types': gtsObjectLiteralTypes,
@@ -62,7 +57,6 @@ const rules: Record<string, Rule.RuleModule> = {
   'no-array-sort-mutation': noArraySortMutation,
   'no-assert-truthiness': noAssertStringOrNumber,
   'no-floating-results': noFloatingVoids,
-  'no-import-workspace-subfolders': noImportSubfolders,
   'no-jest-to-be': noJestToBe,
   'no-react-hook-mutation-dependency': noReactHookMutationDependency,
 } as unknown as Record<string, Rule.RuleModule>;
