+++ libs/eslint-plugin-vx/src/util/BUILD.bazel
@@ -1,0 +1,10 @@
+load("//tools/ts_build:ts_library.bzl", "ts_library")
+
+ts_library(
+    name = "util",
+    deps = [
+        "//:node_modules/@types/node",
+        "//:node_modules/@typescript-eslint/utils",
+        "//:node_modules/typescript",
+    ],
+)
