+++ libs/eslint-plugin-vx/tests/rules/BUILD.bazel
@@ -8,8 +8,7 @@ ts_tests(
         "//:node_modules/@typescript-eslint/parser",
         "//:node_modules/@typescript-eslint/utils",
         "//:node_modules/react",
-        "//:node_modules/typescript",
-        "//libs/eslint-plugin-vx/src",
+        "//libs/eslint-plugin-vx/src/rules",
         "//libs/types/src",
     ],
 )
