+++ libs/eslint-plugin-vx/tests/rules/gts_func_style.test.ts
@@ -6,7 +6,7 @@ const ruleTester = new RuleTester({
   parserOptions: {
     ecmaVersion: 2018,
     tsconfigRootDir: join(__dirname, '../fixtures'),
-    project: './tsconfig.json',
+    project: './tsconfig.fixtures.json',
   },
   parser: require.resolve('@typescript-eslint/parser'),
 });
