+++ package.json
@@ -73,17 +73,6 @@
     "dotenv-expand": "9.0.0",
     "esbuild": "0.21.2",
     "esbuild-runner": "2.2.2",
-    "eslint-config-airbnb": "19.0.4",
-    "eslint-config-airbnb-base": "15.0.0",
-    "eslint-config-prettier": "9.0.0",
-    "eslint-import-resolver-node": "0.3.9",
-    "eslint-import-resolver-typescript": "3.6.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jest": "27.9.0",
-    "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-n": "17.10.2",
-    "eslint-plugin-prettier": "5.0.0",
-    "eslint-plugin-react": "7.31.8",
     "events": "3.3.0",
     "express": "4.18.2",
     "fast-check": "2.23.2",
@@ -180,6 +169,7 @@
     "zod": "3.23.5"
   },
   "devDependencies": {
+    "@eslint/js": "8.57.1",
     "@jest/transform": "^29.7.0",
     "@jest/types": "29.6.3",
     "@storybook/addon-a11y": "7.2.2",
@@ -216,7 +206,9 @@
     "@types/connect": "3.4.35",
     "@types/debug": "4.1.8",
     "@types/deep-eql": "4.0.2",
-    "@types/eslint": "8.4.1",
+    "@types/eslint": "8.56.12",
+    "@types/eslint-plugin-jsx-a11y": "^6.9.0",
+    "@types/eslint__js": "8.42.3",
     "@types/express": "4.17.14",
     "@types/fast-text-encoding": "1.0.1",
     "@types/fetch-mock": "7.3.3",
@@ -267,22 +259,34 @@
     "@types/uuid": "9.0.5",
     "@types/w3c-web-usb": "1.0.6",
     "@types/yargs": "17.0.22",
-    "@typescript-eslint/eslint-plugin": "6.7.0",
-    "@typescript-eslint/parser": "6.7.0",
-    "@typescript-eslint/rule-tester": "6.21.0",
-    "@typescript-eslint/utils": "6.7.0",
+    "@typescript-eslint/parser": "7.2.0",
+    "@typescript-eslint/rule-tester": "7.2.0",
+    "@typescript-eslint/utils": "7.2.0",
     "@vitejs/plugin-react": "1.3.2",
     "ajv": "8.12.0",
     "ajv-draft-04": "1.0.0",
     "cargo-cp-artifact": "0.1.9",
     "csv-parse": "5.5.0",
-    "eslint": "8.57.0",
+    "eslint": "8.57.1",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-import-resolver-node": "0.3.9",
+    "eslint-import-resolver-typescript": "3.6.0",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-jest": "28.8.3",
+    "eslint-plugin-jsx-a11y": "6.10.0",
+    "eslint-plugin-n": "17.10.2",
+    "eslint-plugin-prettier": "5.2.1",
+    "eslint-plugin-react": "7.37.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-storybook": "0.6.10",
     "expect-type": "0.15.0",
     "fast-glob": "3.2.11",
     "get-contrast": "3.0.0",
     "get-port": "5.1.1",
     "glob": "8.0.3",
+    "globals": "15.10.0",
     "husky": "7.0.0",
     "jest": "29.7.0",
     "jest-cli": "^29.7.0",
@@ -312,11 +316,11 @@
     "supertest": "6.1.1",
     "ts-jest": "29.2.5",
     "type-fest": "0.18.1",
+    "typescript-eslint": "7.2.0",
     "vite": "4.5.0"
   },
   "pnpm": {
     "overrides": {
-      "@types/eslint": "8.4.1",
       "graceful-fs": "^4.2.9",
       "nan": "^2.20.0",
       "typescript": "5.6.2"
